requirements.txt
# Memory OS MVP - 最小依赖
# Python 3.10+

# UI
streamlit>=1.28.0

# 文本处理 + LLM
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1

# 数据库
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# 文档解析（可选，后续支持 docx）
python-docx>=1.1.0

# 环境变量
python-dotenv>=1.0.0

# OpenAI 官方（embedding 等）
openai>=1.12.0
